+++ src/EthernetConnectionHandler.cpp
@@ -16,10 +16,10 @@
    INCLUDE
  ******************************************************************************/
 
-#include "Arduino_ConnectionHandlerDefinitions.h"
+#include "ConnectionHandlerDefinitions.h"
 
 #ifdef BOARD_HAS_ETHERNET /* Only compile if the board has ethernet */
-#include "Arduino_EthernetConnectionHandler.h"
+#include "EthernetConnectionHandler.h"
 
 /******************************************************************************
    CTOR/DTOR
